refactor: split restore transaction types
