+++ Packages/macOS/CmuxTerminal/Sources/CmuxTerminal/Lifecycle/TerminalSurfaceRuntimeSpawnTiming.swift
@@ -1,0 +1,5 @@
+/// When a terminal surface should enter its native Ghostty runtime.
+enum TerminalSurfaceRuntimeSpawnTiming: Equatable, Sendable {
+    case immediate
+    case pacedSessionRestore
+}
